Add checbox and clear completed task methods
