+++ src/modules/addRemoveEdit.js
@@ -3,9 +3,9 @@ if (localStorage.getItem('Task List') === null) {
   localStorage.setItem('Task List', JSON.stringify([]));
 }
 
-let taskList = JSON.parse(localStorage.getItem('Task List'));
+export let taskList = JSON.parse(localStorage.getItem('Task List'));
 
-const updateLocalStorage = () => {
+export const updateLocalStorage = () => {
   localStorage.setItem('Task List', JSON.stringify(taskList));
 };
 
@@ -32,19 +32,19 @@ export const displayTasks = () => {
   }
 };
 
-// function to make use of checkbox
-const checkBox = (checkBoxID) => {
-  const indexID = checkBoxID.getAttribute('id');
-  const inputTask = document.querySelectorAll('.taska');
-  if (checkBoxID.checked) {
-    inputTask[indexID - 1].style.textDecoration = 'line-through';
-    taskList[indexID - 1].completed = true;
-  } else {
-    inputTask[indexID - 1].style.textDecoration = 'none';
-    taskList[indexID - 1].completed = false;
-  }
-  updateLocalStorage();
-};
+// // function to make use of checkbox
+// const checkBox = (checkBoxID) => {
+//   const indexID = checkBoxID.getAttribute('id');
+//   const inputTask = document.querySelectorAll('.taska');
+//   if (checkBoxID.checked) {
+//     inputTask[indexID - 1].style.textDecoration = 'line-through';
+//     taskList[indexID - 1].completed = true;
+//   } else {
+//     inputTask[indexID - 1].style.textDecoration = 'none';
+//     taskList[indexID - 1].completed = false;
+//   }
+//   updateLocalStorage();
+// };
 
 // function to edit task and display delete button
 const taskED = (EDId) => {
@@ -110,23 +110,22 @@ export const refreshList = () => {
   displayTasks();
 };
 
-// To clear completed tasks
-export const clearButton = () => {
-  for (let i = taskList.length - 1; i >= 0; i -= 1) {
-    if (taskList[i].completed) {
-      taskList.splice(i, 1);
-    }
-  }
-  for (let b = 0; b < taskList.length; b += 1) {
-    taskList[b].index = b + 1;
-  }
-  updateLocalStorage();
-  displayTasks();
-};
+// // To clear completed tasks
+// export const clearButton = () => {
+//   for (let i = taskList.length - 1; i >= 0; i -= 1) {
+//     if (taskList[i].completed) {
+//       taskList.splice(i, 1);
+//     }
+//   }
+//   for (let b = 0; b < taskList.length; b += 1) {
+//     taskList[b].index = b + 1;
+//   }
+//   updateLocalStorage();
+//   displayTasks();
+// };
 
 // for onClick functions
 window.deleteTask = deleteTask;
 window.taskED = taskED;
-window.checkBox = checkBox;
 window.updateTask = updateTask;
 window.inputSubmit = inputSubmit;
